Add errors to upload summary
After a user has uploaded a CSV, one of two things will happen:

a) The CSVs are unprocessable and they get redirected to an error page
(currently very ugly just showing a json error)

b) The CSVs were processed by the API and we see the page that Amir
wrote with a summary of all assets extracted from the files, but now it
also has any errors in the table. So far, these do not link to the error
detail pages but they soon will. Note that the asset type is missing
from the table, as an unprocessed item doesn't have a type, and the ID
is used instead of the name to identify it.