Add error detail pages for CSV upload
Now when you click on an error in the upload summary page, you get a
page that gives more detail on the error, including which fields were
the problem if applicable.

I have also added the asset title and type to the errors in the summary
table, wherever they are available.

For this to work, we need to merge the [relevant changes to the API](https://github.com/co-cddo/data-marketplace-api/pull/36)
that add more information to the error responses so these should be
reviewed together