Add catchall error page for truly horrible CSV uplaod
If the CSV files uploaded are so poorly formatted (ie not even parseable
as a CSV) the backend can't even give us a general error message.
Therefore I've just caught every possible error that might occur with
the upload with a single page that says we couldn't parse the files.

This probably isn't the best solution, since it will also catch a
general internal server error that's our fault and blame it on the user,
but it'll do for now. Especially since we don't have a page for
our-fault server errors at all.