Re-add check for errors on upload summary page
At some point, we lost the check for 'hasErrors' in the upload summary
template which decided whether or not we show the publish button as well
as the message shown at the top of the page. This adds it back in, so
that if there are errors, you see a message saying as much at the top of
the page and the publish button is not rendered.